+++ changelog.md
@@ -7,6 +7,7 @@
 ### Improvements
 
 - #876 Changing password in backend via modal dialog.
+- #871 Show the total media file size in dashboard stats.
 
 ### Bugfixes